Introducing BUFR to wradlib
The changeset introduces a wide range of changes which allow the use of BUFR files
in wradib. It introduces a new module named bufr.py which provides a wrapper around
the OPERA BUFR software (version 3.1). The BUFR software library was made part of
the repo in order to allow future developers to add changes. Changes also affect
setup.py because the BUFR library needs to be compiled in the course of installation.
Also added an example script and one example data set. Documentation still needs
to be comlpleted, but will be added soon. Please note the the current type of wrapping
is hopefully preliminary. Due to some not yet resolved problems in wrapping a shared
library using ctypes and distutils, wradlib.bufr.decodebufr simply calls the
executable for BUFR decoding (decbufr), reads and parses the temporary output files
and deletes these output files. This is of course not very efficient becasue it
involves redundant reading and writing. However, until the problem is resolved, this
will be the way wradlib will use the BUFR software.

The changeset also introduces a grid plot for cartesian grids to wradlib.vis.