ENH: the class CartesianVolume does now allow for the creation of Pseudo CAPPIs
However, the solution is preliminary (using a Boolean flag argument named pseudocappi). The question which areas in the volume should
be masked as NaN. If we want to create a Pseudo CAPPI, only the areas beyond the range
should be masked. If we want to create a real volume (one CAPPI or layers of real CAPPIs)
also the blind areas (e.g. cone of silence and altitude below the lowest elevation)
should be masked. Maybe we should add one base class and then inherit CAPPI and Pseudo CAPPI
classes.