FIX: setup.py shjould also work with the install --user option.
All files from the directories which are *not* the package (i.e. wradlib)
directories are considered *data* files. In order to make sure these files
are installed to the same site-package directory as the wradlib module files, we
consider the "--user" which might be provided via "python setup.py install --user".
In this case, the target location for the data files is determined via site.USER_SITE.