FIX: Changed setup.py so that it will not produce an installation error on systems other than Windows and Linux.
BUFR compilation is not yet available on Mac OS X from etup.py. Until we find a way to achieve
this, the installation should only give a warning, but should complete without
an error even on Mac. For now, setup.py will copy a dummy README.txt file to the wradlib/bufr_3.1
directory which states that BUFR compilation is not available on Mac, yet.