FIX: Removed PyTables dependency in wradlib.io
PyTables was used to allow convenience import/export of data arrays to/from
hdf5 files. This was implemented via wradlib.io.to_hdf5 and wradlib.io.from_hdf5.
However, PyTables tends to conflict particularly with the netCDF4 module, probably
due to conflicting hdf5 DLLs. So I changed wradlib.io.to_hdf5 and wradlib.io.from_hdf5
so that they use h5py instead of PyTables. This might cause a little bit of loss
in compression efficiency and speed, but the conflicts where quite unpredictable
and version dependent, so unfortunately, we had to get rid of PyTables. Call signature
of the two conveneince functions stays the same, so users should not notice the change.