ENH: Added new keyword argument "show" to vis.polar_plot2 in order to allow multiple subplots.
The default value of show is True. If show is not explicitely passed, the behaviour
of the funciton remains unchanged (i.e. the plot is immediately shown). If show
is set to False, the figure is not plotted and further subplots can be added using
the fig and aspos arguments of polar_plot2.