REF: xvalidate method in adjust now returns NaN values for invalid data points.
In the previous version, the xvalidate method only returned pairs of observed and
estimated if both observed and estimated were valid values. This had a major
disadvantage: You could not relate the values of the returned obs/est pairs to
the original rain gage locations, because there was no mechanism for IDing and
the length of the output arrays depended on the number of valid obs/est pairs.

Now, the returned obs/est pairs have the same length as the original input of
obs. If obs or est have invalid values, the corresponding entries are NaN. This way,
we can relate the output to the exact rain gage locations via indexing.

Basically, returning obs is redundant (because it is the same as the input obs).
But we keep it this way in order to reduce the effect of this change to the API.