MIN: Making the use of "mean" the default to compute MFB, and  "median" optional.
Using AdjustMFB was using the median to compute the mean field bias in order
to make the correction more robust against outliers. In order to account for the
name (MEAN field bias), the behaviour was changed. Now, the field bias is, by default,
computed using the mean of all obs/est ratios. By using the argument "biasby" when
calling the AdjustMFB object, you can decide to use the median (biasby="median") instead.