ENH: Changed the way PhiDP values are filled in function fill_phidp in order to avoid Kdp artefacts.
Citing from the new function docstring: Contiguous NaN regions are filled by the average
of the values of the margins that surround the NaN region. At the left and
right margins of the array, these averages are extrapolated to the end. Using the
average over the margins takes into account noisy PhiDP.

As a consequence, a contiguous region of missing PhiDP will be filled by constant
values determined by the edges of that region. Thus, the derivative (i. e. Kdp) in
that region will be zero. This bahaviour is more desirable than the behaviour
produced by linear interpolation because this will cause arbitrary Kdp values
in case of noisy PhiDP profiles with large portions of missing data.