ENH: Making Mean Field Bias Adjustment (MFB) more robust by using linear regression.
If biasby argument to MFB.__call__ was "mean", the bias was computed by simply
computing the mean of the ratios between obs and est. Now a new argument value is
introduced which is "linregr" (now the default value). In that case, the bias is
computed from the inverse slope of a linear regression forced through (0,0).