MIN: Introducing exception handling for reading input files in read_EDGE_netcdf.
The aim is to at least close files if an error occus during reading,
using try except finally. This might be a model for all input functions. The
exception will be raised (or can be caught outside the function), but the file will
be closed if it was successfully opened.