ENH: Added function aggregate_equidistant_tseries to util module.
This function aggregates time series data under the assumption that the source
and the target time series are equidistant, and that the source time steps
regularly fit into the target time steps (no shifts at the boundaries).
This is the most trivial aggregation scenario.

However, the function allows for gaps in the source data. This means, we assume the
data to be equidistant (each item represents a time step with the same length),
but it does not need to be contiguous. NaN values in the source data are allowed,
too.

The output, though, will be a contiguous time series. This series will have NaN
values for those target time steps which were not sufficiently supported by source
data. The decision whether the support was "sufficient" is based on the argument
*minpercvalid*. This argument specifies the minimum percentage of valid source
time steps inside one target time step (valid meaning not NaN and not missing) which
is required to compute an aggregate. The default value of minpercvalid is 100 percent.
This means no gaps are allowed, and a target time step will be NaN if only one source
time step is missing.

Aggregation methods at the moment are "sum" and "mean" of the source data.