FIX: removed assertion to avoid errors with new DX-files.
Apparently the DX-format writers at the DWD have a strange way to calculate
product length. It looks like the 0x03 character used to separate the header
from the data is not counted, or counted once too many or, I don't know.
Fact is, you can have every combination of even and odd byte length header and
even or odd byte length data section.
However the data must be even length and apparently it doesn't hurt to
read one byte less than implied by product-length - header-length, so that's
going to be the way we move forward.