MIN: Modification of aggregate_equidistant_tseries in module util.
The function was changed in a way that we now check for each target time interval
whether source and target time specs are consistent. If they are not, the target value
for that target interval will be NaN. Bfore that change, the entire target series
was set to NaN if the specs for only one target interval was inconsistent.