ENH: refined optional_import to allow global imports.
function now always returns an object (the actual module on success,
a stub object on failure), so that imports may also be done globally
without breaking the code on a failed import. The lazy import will be
revealed as soon as any attribute is requested from the stub object.

This way failed optional imports won't clutter your command window with
warnings but will give you a clear error message, once you try to access
the contents of the missing module.