OPT: Changed bufr module so it can include the BUFR software from any directory.
As the BUFR software is no longer included in the wradlib package, users have to
download it from the OPERA website and then extract the archives (executables plus
tables) to any directory. Users need to create an environment variable BUFRLIB
which points to that directory.

Also the bufr module has been adapted to BUFR version 3.2 which used different
fixed width output tables (UGLY).

Beyond, users are informed on potential sources fo failure of the bufr module.