FIX: Removed the specific version of Sphinx in the requirements.
Fixing the version of Sphinx in requirements.txt (as done in revision 690)
was probably not a good idea. The current Sphinx version seems to cause problems
on Codeship, but has not been reported to cause problems elsewhere.

Dpendency checking in setup.py as it is now does not allow to assert a maximum
version number. Before constraining Sphinx versions we would need to revise setup.py,
but right now, the change in revision 690 did more harm than help.

On Codeship, we will continue to install Sphinx before the requirements are processed.