ENH: Migrating module conversion from PyRadarMet to wradlib module trafo.
This is is kind of a template for migrating code from the GitHub repository
nguy/PyRadarMet to wradlib/wradlib. In this case, simple conversion functions
for wind speed are migrated from PyRadarMet.conversion to wradlib.trafo. Here,
we also examplify wradlib's testing strategies: We use doctests as well as
one unittest per function.

Please also note that we did not migrate the PyRadarMet functions ``dBZ2Z`` and
``Z2dBZ`` because these do the same as ``wradlib.trafo.idecibel`` and ``wradlib.trafo.decibel``,
respectively, so there would be a conflict.

Please also note that I had to rewrite the docstrings to make them compatible with wradlib.
In addition, I renamed the function arguments to achieve some kind of harmonization.
However, that should not change the behaviour as comopared to PyRadarMet.