REF: Revision of the adjust module with regard to AdjustMFB and overall documentation.
Parameters that control the behaviour of AdjustMFB are passed during initialisation
of the adjustmet object via a dictionary mfb_args. This way, the usage of AdjustMFB
becomes consistent with the other adjustment methods, meaning: all control parameters
are passed during initialisation. Calling the object is the same for all adjustment classes.

Documentation has been revised to make the behaviour more transparent. Documentation
of parameters and examples has been moved to AdjustBase. The docstrings of all
adjustment classes are pointing there. This way, redundancy is minimized.