Changes in # read metadata under get_radiosonde()
A bug in the original code throws an error when the station being loaded with get_radiosonde() does not have the 'Station identifier' parameter. The problem arises because an iterator of parameters is created and then some transformations are made considering the indexing of the elements of the iterator, under the # read metadata section. When the 'Station identifier' parameter (which is the first element) is missing, the wrong parameter is selected and the error occurs. In this commit, changes in the # read metadata section allow for transforming the set of parameters of metadata without referring to the index, in order to avoid the problem of not having the 'Station identifier' parameter.