added 3D handling
xyz is an array of cartesian coordinates for every spherical coordinate, so the possible dimensions are (elevations, azimuths, range, 3). For 2d, it either has (elevations, range, 3) or (azimuths, range, 3) dimensions. For 3d, the only option is the full (elevations, azimuths, range, 3) dimensions. Thus, i think adding this two lines for the 3d case will not break other functionalities.